// sourcecode/BTrak.Packages/snova-rostering/module/module-components/src/lib/rostering/models/employee-model.ts
import { SearchCriteriaInputModelBase } from "./searchCriteriaInputModelBase";

export class EmployeeListModel extends SearchCriteriaInputModelBase {
    employmentStatusId: string;
    employmentStatusName: string;
    departmentId: string;
    designationId: string;
    designationName: string;
    employeeId: string;
    fullName: string;
    userId: string;
    employeeNumber: string;
    genderId: string;
    gender: string;
    maritalStatusId: string;
    maritalStatus: string;
    nationalityId: string;
    nationality: string;
    dateOfBirth: Date;
    smoker: boolean;
    militaryService: boolean;
    nickName: string;
    taxCode: string;
    branchId: string;
    entityId : string;
    branchName: string;
    roleIds: any[];
    permittedBranches:any;
    roleName: string;
    timeZoneId: string;
    timeZoneName: string;
    mobileNo: string;
    isTerminated: boolean;
    isActive: boolean;
    isActiveOnMobile: boolean;
    firstName: string;
    surName: string;
    userName: string;
    email: string;
    profileImage: string;
    createdDateTime: Date;
    lastConnection: Date;
    registeredDateTime: Date;
    timeStamp: any;
    totalCount: number;
    isArchived: boolean;
    emailSearchText: string;
    lineManagerId: string;
    jobCategoryId: string;
    dateOfJoining: Date;
    isUpsertEmployee: boolean;
    jobCategoryType: string;
    shiftTimingId: string;
    SpName: string;
    pageNo: number;
    pageSize: number;
    macAddress: string;
    employeeShiftId: string;
    ipNumber: string;
}